refactor(docs): extract _copy_dataset_summary into a Sphinx extension
Moves _copy_dataset_summary (event: build-finished) from conf.py into
docs/source/_extensions/copy_dataset_summary.py. Copies the catalog
dataset_summary.csv shipped inside the eegdash package into the build's
_static directory so the dataset_summary page and JS counters can fetch it.
Helpers moved alongside: none (self-contained). Drops the now-unused
`shutil` import from conf.py. conf.py shrinks from 4096 to 4076 LOC.

Part of C1 in docs_pipeline_architecture_review.md
